requirements.txt
# for drawing text/images
pillow==10.1.0
# pint for unit conversion
pint==0.22
# for drawing little map
geotiler==0.15
# for loading GPX files
gpxpy==1.6.1
# for loading FIT files
fitdecode==0.10.0
# for lat/lon/speed calcs
geographiclib==1.52
progressbar2==4.2.0
requests==2.31.0
pycairo==1.23.0
sqlitedict==2.1.0

fitparse==1.2.0
fit-tool==0.9.13
gpxpy==1.6.2